one class per file refactoring
