+++ packages/tasit-action/src/ProviderFactory.js
@@ -1,0 +1,88 @@
+import "ethers/dist/shims.js";
+// Note: ethers SHOULD be imported from their main object
+// shims aren't injected with package import
+import { ethers } from "ethers";
+const config = require("config");
+
+export class ProviderFactory {
+  static getProvider = () => {
+    const { provider } = config;
+    const json = provider;
+    return ProviderFactory.createProvider(json);
+  };
+
+  static getDefaultConfig = () => {
+    return {
+      network: "mainnet",
+      provider: "fallback",
+      pollingInterval: 4000,
+      jsonRpc: {
+        url: "http://localhost",
+        port: 8545,
+        allowInsecure: false,
+      },
+    };
+  };
+
+  static createProvider = ({
+    network,
+    provider,
+    pollingInterval,
+    jsonRpc,
+    infura,
+    etherscan,
+  }) => {
+    const networks = ["mainnet", "rinkeby", "ropsten", "kovan", "other"];
+    const providers = ["default", "infura", "etherscan", "jsonrpc"];
+
+    if (!networks.includes(network)) {
+      throw new Error(`Invalid network, use: [${networks}].`);
+    }
+
+    if (!providers.includes(provider)) {
+      throw new Error(`Invalid provider, use: [${providers}].`);
+    }
+
+    if (provider === "fallback") network = "default";
+    if (network === "mainnet") network = "homestead";
+    else if (network === "other") network = undefined;
+
+    const defaultConfig = ProviderFactory.getDefaultConfig();
+
+    let ethersProvider;
+
+    switch (provider) {
+      case "default":
+        ethersProvider = ethers.getDefaultProvider(network);
+
+      case "infura":
+        ethersProvider = new ethers.providers.InfuraProvider(
+          network,
+          infura.apiKey
+        );
+
+      case "etherscan":
+        ethersProvider = new ethers.providers.EtherscanProvider(
+          network,
+          etherscan.apiKey
+        );
+
+      case "jsonrpc":
+        let { url, port, user, password, allowInsecure } = jsonRpc;
+        if (url === undefined) url = defaultConfig.jsonRpc.url;
+        if (port === undefined) port = defaultConfig.jsonRpc.port;
+        if (allowInsecure === undefined)
+          allowInsecure = defaultConfig.jsonRpc.allowInsecure;
+
+        ethersProvider = new ethers.providers.JsonRpcProvider(
+          { url: `${url}:${port}`, user, password, allowInsecure },
+          network
+        );
+    }
+
+    if (pollingInterval) ethersProvider.pollingInterval = pollingInterval;
+    return ethersProvider;
+  };
+}
+
+export default ProviderFactory;
